build: creating SD folder devicetree fix
If buildroot does not generate devicetree.dtb file, it is used the one
from binaries/

+++ build/create_sd_files.sh
@@ -5,6 +5,7 @@ SD_folder="sd_card"
 filesystem="rootfs.cpio.uboot"
 kernel="uImage"
 devicetree="zynq-zybo.dtb"
+devicetreeCompiled="binaries/devicetree.dtb"
 buildroot_output="buildroot/output/images/"
 
 script_name=`basename $0`
@@ -29,9 +30,8 @@ then
 	echo "Please cd to $expected_folder and re-run $script_name"
 else
 	#check files exist
-	if [ ! -f $buildroot_output$filesystem ] || 
-		[ ! -f $buildroot_output$kernel ] || 
-		[ ! -f $buildroot_output$devicetree ];
+	if [ ! -f $buildroot_output$filesystem ] ||
+		[ ! -f $buildroot_output$kernel ];
 	then
 		echo "Some files are missing. Did you forget to run buildroot?"
 	else #copy and rename files
@@ -42,8 +42,15 @@ else
 			mkdir $SD_folder
 		fi
 
+		if [ ! -f $buildroot_output$devicetree ]
+		then
+			echo "Devicetree was not generated with buildroot"			
+			echo "Using device-tree.dtb from binaries/ folder"
+			cp $devicetreeCompiled $SD_folder"/devicetree.dtb"
+		else
+			cp $buildroot_output$devicetree $SD_folder"/devicetree.dtb"
+		fi
 		cp $buildroot_output$filesystem $SD_folder"/uramdisk.image.gz"
 		cp $buildroot_output$kernel $SD_folder"/uImage"
-		cp $buildroot_output$devicetree $SD_folder"/devicetree.dtb"
 	fi
 fi
